+++ chat2db-client/.umirc.prod.desktop.ts
@@ -1,7 +1,8 @@
-import { formatDate } from './src/utils/date';
+import { extractYarnConfig } from './src/utils/webpack';
 import { defineConfig } from 'umi';
 const MonacoWebpackPlugin = require('monaco-editor-webpack-plugin');
-// const UMI_PublicPath = process.env.UMI_PublicPath || './';
+
+const yarn_config = extractYarnConfig(process.argv);
 
 const chainWebpack = (config: any, { webpack }: any) => {
   config.plugin('monaco-editor').use(MonacoWebpackPlugin, [
@@ -17,4 +18,15 @@ export default defineConfig({
   define: {
     'process.env.UMI_ENV': process.env.UMI_ENV,
   },
+  headScripts: [
+    `window.dataLayer = window.dataLayer || [];
+    function gtag() {
+      window.dataLayer.push(arguments);
+    }
+    gtag('js', new Date());
+    gtag('config', 'G-V8M4E5SF61', {
+      platform: 'DESKTOP',
+      version: '${yarn_config['app_version']}'
+    });`,
+  ],
 });