+++ chat2db-client/src/components/SearchResult/TableBox.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useMemo, useState } from 'react';
 import { TableDataType } from '@/constants/table';
-import { IManageResultData, ITableHeaderItem } from '@/typings/database';
+import { IManageResultData, IResultConfig, ITableHeaderItem } from '@/typings/database';
 import { formatDate } from '@/utils/date';
-import { Button, message, Modal, Table } from 'antd';
+import { Button, message, Modal, Pagination, Select, Table } from 'antd';
 import antd from 'antd';
 import { BaseTable, ArtColumn, useTablePipeline, features, SortItem, BaseTableProps } from 'ali-react-table';
 import Iconfont from '../Iconfont';
@@ -15,11 +15,14 @@ import styles from './TableBox.less';
 import { ThemeType } from '@/constants';
 import i18n from '@/i18n';
 import { compareStrings } from '@/utils/sort';
+import MyPagination from './Pagination';
 
 interface ITableProps {
   className?: string;
   data: IManageResultData;
-  isLoading?: boolean;
+  config: IResultConfig;
+  onConfigChange: (config: IResultConfig) => void;
+  onSearchTotal: () => Promise<number | undefined>;
 }
 
 interface IViewTableCellData {
@@ -43,7 +46,7 @@ const DarkSupportBaseTable: any = styled(BaseTable)`
 `;
 
 export default function TableBox(props: ITableProps) {
-  const { className, data, isLoading } = props;
+  const { className, data, config, onConfigChange, onSearchTotal } = props;
   const { headerList, dataList, duration, description } = data || {};
   const [viewTableCellData, setViewTableCellData] = useState<IViewTableCellData | null>(null);
   const [appTheme] = useTheme();
@@ -75,13 +78,29 @@ export default function TableBox(props: ITableProps) {
     () =>
       (headerList || []).map((item, index) => {
         const { dataType, name } = item;
-        const isFirstLine = index === 0;
         const isNumber = dataType === TableDataType.NUMERIC;
+        const isNumericalOrder = dataType === TableDataType.CHAT2DB_ROW_NUMBER;
+        if (isNumericalOrder) {
+          return {
+            code: 'No.',
+            name: 'No.',
+            key: name,
+            lock: true,
+            width: 48,
+            features: { sortable: compareStrings },
+            render: (value: any) => {
+              return (
+                <div className={styles.tableItem}>
+                  <div>{value}</div>
+                </div>
+              );
+            },
+          };
+        }
         return {
           code: name,
           name: name,
           key: name,
-          lock: isFirstLine,
           width: 120,
           render: (value: any, row: any, rowIndex: number) => {
             return (
@@ -94,6 +113,7 @@ export default function TableBox(props: ITableProps) {
               </div>
             );
           },
+          // 如果是数字类型，因为后端返回的都是字符串，所以需要调用字符串对比函数来判断
           features: { sortable: isNumber ? compareStrings : true },
         };
       }),
@@ -111,7 +131,7 @@ export default function TableBox(props: ITableProps) {
           if (type === TableDataType.DATETIME && i) {
             rowData[columns[index].name] = formatDate(i, 'yyyy-MM-dd hh:mm:ss');
           } else if (i === null) {
-            rowData[columns[index].name] = '[null]';
+            rowData[columns[index].name] = '';
           } else {
             rowData[columns[index].name] = i;
           }
@@ -144,20 +164,44 @@ export default function TableBox(props: ITableProps) {
       }),
     );
 
+  const onPageNoChange = (pageNo: number) => {
+    onConfigChange && onConfigChange({ ...config, pageNo });
+  };
+  const onPageSizeChange = (pageSize: number) => {
+    onConfigChange && onConfigChange({ ...config, pageSize, pageNo: 1 });
+  };
+
+  const onClickTotalBtn = async () => {
+    if (props.onSearchTotal) {
+      return await props.onSearchTotal();
+    }
+  };
   return (
     <div className={classnames(className, styles.tableBox)}>
       {columns.length ? (
-        // <Table pagination={false} columns={columns} dataSource={tableData} scroll={{ y: '100vh' }} size="small" />
         <>
+          <div className={styles.toolBar}>
+            <div className={styles.toolBarItem}>
+              <MyPagination
+                data={config}
+                onPageNoChange={onPageNoChange}
+                onPageSizeChange={onPageSizeChange}
+                onClickTotalBtn={onClickTotalBtn}
+              />
+            </div>
+          </div>
           <DarkSupportBaseTable
             className={classnames({ dark: isDarkTheme }, props.className, styles.table)}
             components={{ EmptyContent: () => <h2>{i18n('common.text.noData')}</h2> }}
-            isLoading={isLoading}
+            isStickyHead
+            stickyTop={31}
             {...pipeline.getProps()}
           />
-          <div className={styles.statusBar}>{`${i18n('common.text.result')}：${description}. ${i18n(
-            'common.text.timeConsuming',
-          )}：${duration}ms`}</div>
+          <div className={styles.statusBar}>
+            <span>{`【${i18n('common.text.result')}】${description}.`}</span>
+            <span>{`【${i18n('common.text.timeConsuming')}】${duration}ms.`}</span>
+            <span>{`【${i18n('common.text.searchRow')}】${tableData.length} ${i18n('common.text.row')}.`}</span>
+          </div>
         </>
       ) : (
         <StateIndicator state="success" text={i18n('common.text.successfulExecution')} />