+++ chat2db-client/src/components/Tabs/index.tsx
@@ -117,6 +117,5 @@ export default memo<IProps>(function Tab(props) {
         </div>
       </div>
     }
-
   </div >
 })