+++ chat2db-client/src/i18n/zh-cn/workspace.ts
@@ -8,4 +8,6 @@ export default {
   'workspace.menu.unPin': '取消置顶',
   'workspace.menu.deleteTablePlaceHolder': '请输入你要删除的表名',
   'workspace.tips.affirmDeleteTable': '输入的表名与要删除的表名不一致，请再次确认',
-}
+  'workspace.table.total': '总数',
+  'workspace.table.total.tip': '加载总行数',
+};