+++ chat2db-client/src/main/utils.js
@@ -1,6 +1,8 @@
 const { DEV_WEB_URL } = require('./constants');
 const path = require('path');
 const url = require('url');
+const { version } = require('os');
+const fs = require('fs');
 
 /**
  * 加载主进程前端资源
@@ -10,13 +12,12 @@ function loadMainResource(mainWindow) {
   if (process.env.NODE_ENV === 'development') {
     mainWindow.loadURL(DEV_WEB_URL);
     mainWindow.webContents.openDevTools();
-
     // 监听应用程序根路径下的所有文件，当文件发生修改时，自动刷新应用程序
     require('electron-reload')(path.join(__dirname, '..'));
   } else {
     mainWindow.loadURL(
       url.format({
-        pathname: path.join(__dirname, '../..', './dist/index.html'),
+        pathname: path.join(__dirname, '../..', `./versions/${readVersion()}`, `./dist/index.html`),
         protocol: 'file:',
         slashes: true,
       }),
@@ -24,6 +25,19 @@ function loadMainResource(mainWindow) {
   }
 }
 
+/**
+ * 应该读取哪个版本的资源
+ * @param {*} 
+ */
+function readVersion() {
+  let version = '';
+  if (process.env.NODE_ENV !== 'development') {
+    version = fs.readFileSync(path.join(__dirname, '../..', './versions/version'));
+  }
+  return version.toString().trim();
+}
+
 module.exports = {
   loadMainResource,
+  readVersion
 };