+++ chat2db-client/src/typings/database.ts
@@ -18,7 +18,18 @@ export interface IManageResultData {
   description: string;
   message: string;
   sql: string;
+  originalSql: string;
   success: boolean;
   uuid?: string;
   duration: number;
+  fuzzyTotal: string;
+  hasNextPage: boolean;
+}
+
+/** 查询结果 配置属性 */
+export interface IResultConfig {
+  pageNo: number;
+  pageSize: number;
+  total: number | string;
+  hasNextPage: boolean;
 }