+++ chat2db-server/chat2db-server-domain/chat2db-server-domain-core/src/main/java/ai/chat2db/server/domain/core/impl/ViewServiceImpl.java
@@ -11,6 +11,6 @@ public class ViewServiceImpl implements ViewService {
 
     @Override
     public ListResult<Table> views(String databaseName, String schemaName) {
-        return ListResult.of(Chat2DBContext.getMetaData().views(databaseName, schemaName));
+        return ListResult.of(Chat2DBContext.getMetaData().views(Chat2DBContext.getConnection(),databaseName, schemaName));
     }
 }