+++ chat2db-server/chat2db-spi/src/main/java/ai/chat2db/spi/sql/SQLExecutor.java
@@ -12,8 +12,10 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import ai.chat2db.server.tools.base.constant.EasyToolsConstant;
+import ai.chat2db.server.tools.common.util.I18nUtils;
 import ai.chat2db.spi.model.*;
 
+import ai.chat2db.spi.util.ResultSetUtils;
 import cn.hutool.core.date.TimeInterval;
 import com.google.common.collect.Lists;
 import lombok.extern.slf4j.Slf4j;
@@ -48,9 +50,9 @@ public class SQLExecutor {
         return INSTANCE;
     }
 
-    public Connection getConnection() throws SQLException {
-        return Chat2DBContext.getConnection();
-    }
+    //public Connection connection throws SQLException {
+    //    return Chat2DBContext.connection;
+    //}
 
     public void close() {
     }
@@ -58,17 +60,18 @@ public class SQLExecutor {
     /**
      * 执行sql
      *
+     * @param connection
      * @param sql
      * @param function
      * @return
      */
 
-    public <R> R executeSql(String sql, Function<ResultSet, R> function) {
+    public <R> R executeSql(Connection connection, String sql, Function<ResultSet, R> function) {
         if (StringUtils.isEmpty(sql)) {
             return null;
         }
         log.info("execute:{}", sql);
-        try (Statement stmt = getConnection().createStatement();) {
+        try (Statement stmt = connection.createStatement();) {
             boolean query = stmt.execute(sql);
             // 代表是查询
             if (query) {
@@ -98,7 +101,7 @@ public class SQLExecutor {
             stmt.setFetchSize(EasyToolsConstant.MAX_PAGE_SIZE);
             TimeInterval timeInterval = new TimeInterval();
             boolean query = stmt.execute(sql.replaceFirst(";", ""));
-            executeResult.setDescription("执行成功");
+            executeResult.setDescription(I18nUtils.getMessage("sqlResult.success"));
             // 代表是查询
             if (query) {
                 ResultSet rs = null;
@@ -115,7 +118,7 @@ public class SQLExecutor {
                         headerList.add(Header.builder()
                             .dataType(ai.chat2db.spi.util.JdbcUtils.resolveDataType(
                                 resultSetMetaData.getColumnTypeName(i), resultSetMetaData.getColumnType(i)).getCode())
-                            .name(resultSetMetaData.getColumnName(i))
+                            .name(ResultSetUtils.getColumnName(resultSetMetaData, i))
                             .build());
                     }
 
@@ -148,22 +151,24 @@ public class SQLExecutor {
     /**
      * 执行sql
      *
+     * @param connection
      * @param sql
      * @return
      * @throws SQLException
      */
-    public ExecuteResult execute(String sql) throws SQLException {
-        return execute(sql, getConnection());
+    public ExecuteResult execute(Connection connection, String sql) throws SQLException {
+        return execute(sql, connection);
     }
 
     /**
      * 获取所有的数据库
      *
+     * @param connection
      * @return
      */
-    public List<String> databases() {
+    public List<String> databases(Connection connection) {
         List<String> tables = Lists.newArrayList();
-        try (ResultSet resultSet = getConnection().getMetaData().getCatalogs();) {
+        try (ResultSet resultSet = connection.getMetaData().getCatalogs();) {
             if (resultSet != null) {
                 while (resultSet.next()) {
                     tables.add(resultSet.getString("TABLE_CAT"));
@@ -178,14 +183,15 @@ public class SQLExecutor {
     /**
      * 获取所有的schema
      *
+     * @param connection
      * @param databaseName
      * @param schemaName
      * @return
      */
-    public List<Map<String, String>> schemas(String databaseName, String schemaName) {
+    public List<Map<String, String>> schemas(Connection connection, String databaseName, String schemaName) {
         List<Map<String, String>> schemaList = Lists.newArrayList();
         if (StringUtils.isEmpty(databaseName) && StringUtils.isEmpty(schemaName)) {
-            try (ResultSet resultSet = getConnection().getMetaData().getSchemas()) {
+            try (ResultSet resultSet = connection.getMetaData().getSchemas()) {
                 if (resultSet != null) {
                     while (resultSet.next()) {
                         Map<String, String> map = new HashMap<>();
@@ -199,7 +205,7 @@ public class SQLExecutor {
             }
             return schemaList;
         }
-        try (ResultSet resultSet = getConnection().getMetaData().getSchemas(databaseName, schemaName)) {
+        try (ResultSet resultSet = connection.getMetaData().getSchemas(databaseName, schemaName)) {
             if (resultSet != null) {
                 while (resultSet.next()) {
                     Map<String, String> map = new HashMap<>();
@@ -217,16 +223,18 @@ public class SQLExecutor {
     /**
      * 获取所有的数据库表
      *
+     * @param connection
      * @param databaseName
      * @param schemaName
      * @param tableName
      * @param types
      * @return
      */
-    public List<Table> tables(String databaseName, String schemaName, String tableName, String types[]) {
+    public List<Table> tables(Connection connection, String databaseName, String schemaName, String tableName,
+        String types[]) {
         List<Table> tables = Lists.newArrayList();
         int n = 0;
-        try (ResultSet resultSet = getConnection().getMetaData().getTables(databaseName, schemaName, tableName,
+        try (ResultSet resultSet = connection.getMetaData().getTables(databaseName, schemaName, tableName,
             types)) {
             if (resultSet != null) {
                 while (resultSet.next()) {
@@ -246,15 +254,17 @@ public class SQLExecutor {
     /**
      * 获取所有的数据库表列
      *
+     * @param connection
      * @param databaseName
      * @param schemaName
      * @param tableName
      * @param columnName
      * @return
      */
-    public List<TableColumn> columns(String databaseName, String schemaName, String tableName, String columnName) {
+    public List<TableColumn> columns(Connection connection, String databaseName, String schemaName, String tableName,
+        String columnName) {
         List<TableColumn> tableColumns = Lists.newArrayList();
-        try (ResultSet resultSet = getConnection().getMetaData().getColumns(databaseName, schemaName, tableName,
+        try (ResultSet resultSet = connection.getMetaData().getColumns(databaseName, schemaName, tableName,
             columnName)) {
             if (resultSet != null) {
                 while (resultSet.next()) {
@@ -270,14 +280,15 @@ public class SQLExecutor {
     /**
      * 获取所有的数据库表索引
      *
+     * @param connection
      * @param databaseName
      * @param schemaName
      * @param tableName
      * @return
      */
-    public List<TableIndex> indexes(String databaseName, String schemaName, String tableName) {
+    public List<TableIndex> indexes(Connection connection, String databaseName, String schemaName, String tableName) {
         List<TableIndex> tableIndices = Lists.newArrayList();
-        try (ResultSet resultSet = getConnection().getMetaData().getIndexInfo(databaseName, schemaName, tableName,
+        try (ResultSet resultSet = connection.getMetaData().getIndexInfo(databaseName, schemaName, tableName,
             false,
             false)) {
             List<TableIndexColumn> tableIndexColumns = Lists.newArrayList();
@@ -308,14 +319,15 @@ public class SQLExecutor {
     /**
      * 获取所有的函数
      *
+     * @param connection
      * @param databaseName
      * @param schemaName
      * @return
      */
-    public List<ai.chat2db.spi.model.Function> functions(String databaseName,
+    public List<ai.chat2db.spi.model.Function> functions(Connection connection, String databaseName,
         String schemaName) {
         List<ai.chat2db.spi.model.Function> functions = Lists.newArrayList();
-        try (ResultSet resultSet = getConnection().getMetaData().getFunctions(databaseName, schemaName, null);) {
+        try (ResultSet resultSet = connection.getMetaData().getFunctions(databaseName, schemaName, null);) {
             while (resultSet != null && resultSet.next()) {
                 functions.add(buildFunction(resultSet));
             }
@@ -328,13 +340,14 @@ public class SQLExecutor {
     /**
      * 获取所有的存储过程
      *
+     * @param connection
      * @param databaseName
      * @param schemaName
      * @return
      */
-    public List<Procedure> procedures(String databaseName, String schemaName) {
+    public List<Procedure> procedures(Connection connection, String databaseName, String schemaName) {
         List<Procedure> procedures = Lists.newArrayList();
-        try (ResultSet resultSet = getConnection().getMetaData().getProcedures(databaseName, schemaName, null)) {
+        try (ResultSet resultSet = connection.getMetaData().getProcedures(databaseName, schemaName, null)) {
             while (resultSet != null && resultSet.next()) {
                 procedures.add(buildProcedure(resultSet));
             }